fix: Get Latitude, Longitude from google API
