+++ README.md
@@ -2,8 +2,7 @@
 
 ## How to run locally 
 1. npm install 
-2. npm i react-router react-router-dom
-3. npm run dev
+2. npm run dev
 
 ### Hosted on AWS Cloudfront URL
 https://d1088ylolbm0n2.cloudfront.net
