+++ package.json
@@ -9,7 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@aws-amplify/ui-react": "^5.0.5",
+    "@aws-amplify/ui-react": "^5.3.1",
     "@react-google-maps/api": "^2.19.2",
     "@stripe/react-stripe-js": "^2.3.0",
     "@stripe/stripe-js": "^2.1.5",
@@ -18,8 +18,8 @@
     "@types/react-dom": "18.2.7",
     "antd": "^5.8.5",
     "autoprefixer": "10.4.14",
-    "aws-amplify": "^5.3.5",
-    "axios": "^1.5.1",
+    "aws-amplify": "^5.3.11",
+    "axios": "^1.5.0",
     "eslint": "8.45.0",
     "eslint-config-next": "13.4.12",
     "first": "^0.0.3",
@@ -27,8 +27,6 @@
     "postcss": "8.4.27",
     "react": "18.2.0",
     "react-dom": "18.2.0",
-    "react-router": "^6.16.0",
-    "react-router-dom": "^6.16.0",
     "tailwindcss": "3.3.3",
     "typescript": "5.1.6"
   }
